Use interrupts for serial transfers
Rewrite the serial driver to use interrupts

+++ Application/ModbusRTU.h
@@ -98,7 +98,7 @@ private:
 	uint8_t  m_InputFrame[MODBUS_MAX_FRAME_LENGTH];
 	uint8_t  m_OutputFrame[MODBUS_MAX_FRAME_LENGTH];
 	uint8_t  m_InputFrameLength;
-	uint16_t m_RxTimeout;
+	uint32_t m_RxTimeout;
 
 	static uint16_t crc16(const uint8_t *nData, uint16_t wLength);
